optimized pattern set
